backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
httpx==0.27.2
cryptography==43.0.1
pydantic==2.9.2
python-multipart==0.0.12
torch
onnxruntime
transformers
python-dotenv
